requirements.txt
absl-py==2.3.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
ConfigArgParse==1.7.1
filelock==3.19.1
Flask==3.1.2
fsspec==2025.9.0
future==1.0.0
grpcio==1.75.1
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
lxml==6.0.1
Markdown==3.9
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
OpenNMT-py==1.2.0
packaging==25.0
pandas==2.3.2
pillow==12.0.0
portalocker==3.2.0
protobuf==5.29.5
pybind11==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
sacrebleu==2.5.1
sacremoses==0.1.1
safetensors==0.6.2
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tokenizers==0.22.0
torch==2.9.0
torchtext==0.4.0
tqdm==4.67.1
transformers==4.56.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3
wheel==0.45.1
